+++ voxygen/src/session.rs
@@ -1274,6 +1274,10 @@ impl PlayState for SessionState {
                         global_state.settings.interface.map_show_trees = map_show_trees;
                         global_state.settings.save_to_file_warn();
                     },
+                    HudEvent::RequestSiteInfo(id) => {
+                        let mut client = self.client.borrow_mut();
+                        client.request_site_economy(id);
+                    },
                     HudEvent::ChangeGamma(new_gamma) => {
                         global_state.settings.graphics.gamma = new_gamma;
                         global_state.settings.save_to_file_warn();